PODS-1387 : non uk partnership address
