+++ test/controllers/register/partnership/PartnershipRegisteredAddressControllerSpec.scala
@@ -17,58 +17,29 @@
 package controllers.register.partnership
 
 import audit.testdoubles.StubSuccessfulAuditService
-import connectors.{FakeUserAnswersCacheConnector, RegistrationConnector}
-import controllers.ControllerSpecBase
+import connectors.FakeUserAnswersCacheConnector
 import controllers.actions._
+import controllers.address.NonUKAddressControllerDataMocks
 import forms.address.NonUKAddressFormProvider
 import identifiers.register.partnership.{PartnershipDetailsId, PartnershipRegisteredAddressId}
 import models._
 import org.scalatest.concurrent.ScalaFutures
 import play.api.data.Form
 import play.api.libs.json.Json
-import play.api.mvc.Call
 import play.api.test.Helpers._
-import uk.gov.hmrc.http.HeaderCarrier
-import utils.countryOptions.CountryOptions
-import utils.{FakeCountryOptions, FakeNavigator}
+import utils.FakeNavigator
 import viewmodels.Message
 import viewmodels.address.ManualAddressViewModel
 import views.html.address.nonukAddress
 
-import scala.concurrent.{ExecutionContext, Future}
-
-class PartnershipRegisteredAddressControllerSpec extends ControllerSpecBase with ScalaFutures {
-
-  def onwardRoute: Call = controllers.routes.IndexController.onPageLoad()
-
-  def countryOptions: CountryOptions = new FakeCountryOptions(environment, frontendAppConfig)
+class PartnershipRegisteredAddressControllerSpec extends NonUKAddressControllerDataMocks with ScalaFutures {
 
   val formProvider = new NonUKAddressFormProvider(countryOptions)
   val form = formProvider("error.country.invalid")
   val fakeAuditService = new StubSuccessfulAuditService()
   private val partnershipName = "Test Partnership Name"
 
-  private def fakeRegistrationConnector = new RegistrationConnector {
-    override def registerWithIdOrganisation(utr: String, organisation: Organisation, legalStatus: RegistrationLegalStatus)
-                                           (implicit hc: HeaderCarrier, ec: ExecutionContext): Future[OrganizationRegistration] = ???
-
-    override def registerWithNoIdOrganisation(name: String, address: Address)
-                                             (implicit hc: HeaderCarrier, ec: ExecutionContext): Future[RegistrationInfo] = Future.successful(
-      RegistrationInfo(
-        RegistrationLegalStatus.LimitedCompany,
-        "test-sap-number",
-        false,
-        RegistrationCustomerType.NonUK,
-        None,
-        None
-      )
-    )
-
-    override def registerWithIdIndividual(nino: String)(
-      implicit hc: HeaderCarrier, ec: ExecutionContext): Future[IndividualRegistration] = ???
-  }
-
-  def controller(dataRetrievalAction: DataRetrievalAction = getCompany) =
+  def controller(dataRetrievalAction: DataRetrievalAction = getPartnership) =
     new PartnershipRegisteredAddressController(
       frontendAppConfig,
       messagesApi,
